requirements.txt
flask
google-ads
gunicorn
